+++ im-model/src/main/java/com/zj/model/mod/MessageBean.java
@@ -8,7 +8,7 @@ public class MessageBean {
     public boolean deleted; // false,
     public String content; // null,
     public String updated; // "2020-01-03T09:38:11Z",
-    public int uid; // 4,
+    public long uid; // 4,
     public String created; // "2020-01-03T09:38:11Z",
     public long id = System.currentTimeMillis(); // 991077318262784,
     public int team_id; // 1,
